+++ app/Filament/Resources/Courses/CourseResource.php
@@ -88,7 +88,9 @@ class CourseResource extends Resource
     public static function getRelations(): array
     {
         return [
+            RelationManagers\LessonsRelationManager::class,
             RelationManagers\FinalQuestionsRelationManager::class,
+            RelationManagers\FeedbackRelationManager::class,
         ];
     }